pyproject.toml
[project]
name = "practicepython"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.0",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pyjwt>=2.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "python-multipart>=0.0.20",
    "trio>=0.30.0",
]
